Add files via upload
Bank term deposit marketing supervised machine learning.
The main profit of commercial banks is from the intrest rate difference between deposit collected and loan out. The precondition would be sufficient resource (collected deposits). A bank is trying to learn from the past experience about which group of customers are likely to make term deposits so that the bank could target those customers and gain their deposits before other banks for a competitive advantage.
Used popular classification algorithms Decision Tree, KNN, random forest, boosting, SVM, logistic regression, and Naive Bayes to find the best model for the bank, so that the bank could target and gain the right customers in time to win in the competition. The main stat we focus is the True Positive customers (the customers who will make the deposit as predicted) since the cost of calling is minor but the benefit of correctly predicted customers is high.